feat: reorganize backend
