+++ packages/backend/src/opportunity/domain/spi.ts
@@ -1,7 +1,6 @@
 import type { Maybe, Result } from 'true-myth'
 import { Program } from '../../program/domain/types/types'
 import type { ContactId, OpportunityId, ContactDetails, OpportunityDetails, OpportunityUpdateAttributes, Opportunity } from './types'
-import StatsData from '@tee/common/src/stats/types'
 
 export type ContactRepository = {
   createOrUpdate: (contact: ContactDetails, optIn: true) => Promise<Result<ContactId, Error>>
@@ -15,8 +14,4 @@ export type OpportunityRepository = {
 
 export type MailerService = {
   sendReturnReceipt: (opportunity: Opportunity, program: Program) => Promise<Maybe<Error> | void>
-}
-
-export type StatisticsRepository = {
-  get: () => Promise<Result<StatsData, Error>>
 }
