+++ packages/backend/src/opportunity/infrastructure/api/brevo/brevoDeal.ts
@@ -111,9 +111,10 @@ const getBrevoCreationDates = async (): Promise<Result<Date[], Error>> => {
   const responsePatch = await new BrevoAPI().GetDeals()
 
   if (responsePatch.isOk) {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const brevoDealResponse: BrevoDealResponse = responsePatch.value.data
-    if (brevoDealResponse.items.length === 0) {
+    const brevoDealResponse: BrevoDealResponse = responsePatch.value.data as BrevoDealResponse
+    if (!brevoDealResponse.items) {
+      return Result.err(new Error('No Items field in Brevo API'))
+    } else if (brevoDealResponse.items.length === 0) {
       return Result.err(new Error('Brevo deal list is empty'))
     }
     const dateList: Date[] = []
