+++ packages/backend/src/statistics/application/statisticsService.ts
@@ -1,6 +1,6 @@
 import { Result } from 'true-myth'
 import StatisticsFeatures from '../domain/statisticsFeatures'
-import { brevoRepository } from '../infrastructure/api/brevo/brevoDeal'
+import { brevoRepository } from '../../opportunity/infrastructure/api/brevo/brevoDeal'
 import StatsData from '@tee/common/src/stats/types'
 
 export default class StatisticsService {
