+++ packages/backend/src/statistics/domain/spi.ts
@@ -1,0 +1,6 @@
+import type { Result } from 'true-myth'
+import StatsData from '@tee/common/src/stats/types'
+
+export type StatisticsRepository = {
+  get: () => Promise<Result<StatsData, Error>>
+}
